% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{are-named}
\alias{are-named}
\alias{are_named}
\alias{are_unnamed}
\alias{which_named}
\alias{which_unnamed}
\title{Are elements of a vector named?}
\usage{
are_named(x)

are_unnamed(x)

which_named(x, .p, ...)

which_unnamed(x)
}
\arguments{
\item{x}{vector.}
}
\value{
a logical vector for \verb{are_*} functions. An integer vector for \code{which_} functions.
}
\description{
Are elements of a vector named?
}
