% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_fn.R
\name{as_fn}
\alias{as_fn}
\title{Convert to Function}
\usage{
as_fn(x, ...)
}
\arguments{
\item{x}{A one-sided formula or a function. If \code{.str = TRUE}, it also
accepts a string.}

\item{...}{Must remain empty.}
}
\value{
A function.
}
\description{
This function converts a one-sided formula into a function. It is a
simplified version of \link[rlang:as_function]{rlang::as_function}.
}
