% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sym.R
\name{as_sym}
\alias{as_sym}
\title{Convert Object(s) to Symbol(s)}
\usage{
as_sym(x, .strict = FALSE, .simplify = FALSE, .no_ns = FALSE, .no_args = TRUE)
}
\arguments{
\item{x}{.}

\item{.strict}{.}

\item{.simplify}{.}

\item{.no_ns}{.}

\item{.no_args}{.}
}
\value{
TODO.
}
\description{
Convert Object(s) to Symbol(s)
}
