% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_predicates.R
\name{base-type}
\alias{base-type}
\alias{is_base_type}
\alias{is_base_scalar}
\title{Is an object a base type?}
\usage{
is_base_type(x, class = NULL, type = NULL, n = NULL)

is_base_scalar(x, class = NULL, type = NULL)
}
\description{
Is an object a base type?
}
