% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butlerrr-package.R
\docType{package}
\name{butlerrr-package}
\alias{butlerrr}
\alias{butlerrr-package}
\title{butlerrr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/butlerrr}
  \item \url{https://f-dallolio.github.io/butlerrr/}
  \item Report bugs at \url{https://github.com/f-dallolio/butlerrr/issues}
}

}
\author{
\strong{Maintainer}: Filippo Dall Olio \email{fdallolio@brocku.ca}

}
\keyword{internal}
