% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dots.R
\name{dots-lang}
\alias{dots-lang}
\alias{dots_lang}
\alias{print.dots_lang}
\title{Collect dots as symbolic objects}
\usage{
dots_lang(
  .lang_type = NULL,
  .named = FALSE,
  .name_all = FALSE,
  .dots_env = NULL,
  .out_as_list = FALSE
)

\method{print}{dots_lang}(x, .out_as_list = attr(x, "list_output"))
}
\arguments{
\item{.lang_type}{.}

\item{.named}{.}

\item{.name_all}{.}

\item{.out_as_list}{.}
}
\value{
a list.
}
\description{
Collect dots as symbolic objects
}
