% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sym.R, R/encall.R
\name{encall}
\alias{encall}
\alias{as_syms}
\alias{encalls}
\title{Convert Object(s) to Call(s)}
\usage{
as_syms(
  ...,
  .x = NULL,
  .named = NULL,
  .strict = FALSE,
  .simplify = TRUE,
  .no_ns = FALSE,
  .no_args = FALSE
)

encall(x, .strict = FALSE, .simplify = FALSE, .no_ns = FALSE, .no_args = FALSE)

encalls(
  ...,
  .x = NULL,
  .named = NULL,
  .strict = FALSE,
  .simplify = FALSE,
  .no_ns = FALSE,
  .no_args = FALSE
)
}
\arguments{
\item{...}{.}

\item{.strict}{.}

\item{.simplify}{.}

\item{.no_ns}{.}

\item{.no_args}{.}

\item{x, .x}{.}
}
\value{
TODO.
}
\description{
Convert Object(s) to Call(s)
}
