% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlang.R
\name{enlang}
\alias{enlang}
\alias{enlangs}
\title{Convert to \code{"expressions"} (similar to \code{\link[rlang:expr]{rlang::expr()}})}
\usage{
enlang(x, .strict = FALSE, .simplify = FALSE, .no_ns = FALSE, .no_args = FALSE)

enlangs(
  ...,
  .x = NULL,
  .named = NULL,
  .strict = FALSE,
  .simplify = FALSE,
  .no_ns = FALSE,
  .no_args = FALSE
)
}
\arguments{
\item{x}{.}

\item{.strict}{.}

\item{.simplify}{.}

\item{.no_ns}{.}

\item{.no_args}{.}

\item{...}{.}

\item{.x}{.}

\item{.out_type}{.}
}
\value{
\code{\link[=enlang]{enlang()}} returns a symbolic object. \code{\link[=enlangs]{enlangs()}} returns a list of
symbolic objects.
}
\description{
Convert to \code{"expressions"} (similar to \code{\link[rlang:expr]{rlang::expr()}})
}
