% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fmls.R
\name{fn-fmls}
\alias{fn-fmls}
\alias{fn_fmls}
\alias{fn_fmls_nms}
\alias{fn_fmls_syms}
\alias{fn_fmls_update}
\title{Function Formals}
\usage{
fn_fmls(.fn)

fn_fmls_nms(.fn)

fn_fmls_syms(.fn, .unname_dots = FALSE)

fn_fmls_update(.fn, ...)
}
\arguments{
\item{.fn}{a function.}

\item{.unname_dots}{logical.}

\item{...}{named objects defining the new formals defaults.}
}
\value{
a list (\link{fn_fmls} and \link{fn_fmls_syms}), a character vector (\link{fn_fmls_nms}), or a function(\link{fn_fmls_update}).
}
\description{
Function Formals
}
