% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn-ns}
\alias{fn-ns}
\alias{fn_ns}
\alias{fn_ns_name}
\title{Function Namespace}
\usage{
fn_ns(x, ...)

fn_ns_name(x, ...)
}
\arguments{
\item{x}{A function, list of functions, or a string/character vector with function name(s).}

\item{...}{empty dots.}
}
\value{
The namespace or a list of namespaces of the function in input.
}
\description{
Function Namespace
}
\examples{
fn_ns(mean)
fn_ns("mean")
fn_ns(c("mean", var))
}
