% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_from_chr}
\alias{fn_from_chr}
\title{Function from Name}
\usage{
fn_from_chr(x, ..., stop_not_found = FALSE)
}
\arguments{
\item{x}{A string or a character vector. Also accepts input of the type "ns:::fn".}

\item{...}{Empty dots.}

\item{stop_not_found}{If \code{FALSE} (default) it returns \code{NULL} if the function is not found. An error if \code{TRUE}.}
}
\value{
A function or a list of functions.
}
\description{
Function from Name
}
\examples{
# fn_from_chr("mean")
# fn_from_chr(c("mean", "person"))
# fn_from_chr(c("utils::person", "utils:::person"))
}
