% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn_names.R
\name{function-names}
\alias{function-names}
\alias{fn_name}
\alias{fn_names}
\title{Get function names from functions}
\usage{
fn_name(fn, with_ns = TRUE, strict = FALSE)

fn_names(x, with_ns = TRUE, strict = FALSE)
}
\arguments{
\item{with_ns}{TRUE or FALSE. If TRUE (default), the function name(s) includes the namespace.}

\item{strict}{TRUE or FALSE. If FALSE (default) and `x` is not a function, the function will return `NA`. It will throw an error otherwise.}

\item{x}{a function.}
}
\value{
a string (`fn_name`) or a character vector (`fn_names`).
}
\description{
Get function names from functions
}
