% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos.R
\name{get-pos}
\alias{get-pos}
\alias{get_pos0}
\alias{get_pos}
\alias{get_pos2}
\title{Extract Parts of an Object}
\usage{
get_pos0(x, ..., unname = FALSE)

get_pos(x, ..., simplify_scalar = TRUE, unname = FALSE)

get_pos2(x, ..., unname = TRUE)
}
\arguments{
\item{x}{An R object.}

\item{...}{Integers, string or character vector, or unquoted selection expression. Passed to
\code{\link[tidyselect:eval_select]{tidyselect::eval_select}}.
\itemize{
\item If input is an \strong{unnamed atomic vector}, selection is performed on the input itself.
\item For a \strong{named atomic vector}, set \code{unname = TRUE} to perform selection on the input itself.
}}

\item{unname}{\code{TRUE} or \code{FALSE} (default). If \code{TRUE} and input is a named atomic vector, selection is always performed on the input itself.}

\item{simplify_scalar}{\code{TRUE} (default) or \code{FALSE}. If \code{TRUE} and the output
is a vector of length 1, it returns the output simplified to a scalar.}
}
\value{
The extracted elementsÎ of \code{x}.
}
\description{
Extract Parts of an Object
}
