% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{is-named}
\alias{is-named}
\alias{is_named0}
\alias{is_unnamed0}
\alias{is_named}
\alias{is_unnamed}
\title{Does a vector have names?}
\usage{
is_named0(x)

is_unnamed0(x)

is_named(x, .empty_as_null = FALSE)

is_unnamed(x, .empty_as_null = FALSE)
}
\arguments{
\item{x}{vector.}

\item{.empty_as_null}{\code{TRUE} or \code{FALSE} (default). If \code{TRUE}, a vector with empty names is considered unnamed.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Does a vector have names?
}
