% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_predicates.R
\name{is-symbolic2}
\alias{is-symbolic2}
\alias{is_symbolic}
\alias{is_symbolic2}
\title{Is an object a call or a symbol?}
\usage{
is_symbolic(x)

is_symbolic2(x, all = TRUE)
}
\arguments{
\item{x}{an arbitrary R object.}
}
\value{
TRUE or FALSE or a vector of logicals.
}
\description{
Is an object a call or a symbol?
}
