% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_predicates.R
\name{is_string}
\alias{is_string}
\title{Is object a single string}
\usage{
is_string(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
TRUE if \code{x} is a character of length 1 (i.e. a string) or FALSE.
}
\description{
Is object a single string
}
