% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_predicates.R
\name{is_syntactic_literal}
\alias{is_syntactic_literal}
\title{Is an object a syntactic literal?}
\usage{
is_syntactic_literal(x)
}
\arguments{
\item{x}{an arbitrary R object.}
}
\value{
TRUE or FALSE or a vector of logicals.
}
\description{
Is an object a syntactic literal?
}
