% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_predicates.R
\name{is_whole}
\alias{is_whole}
\title{Is a number whole?}
\usage{
is_whole(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a numeric vector.}

\item{tol}{tolerance.}
}
\value{
TRUE or FALSE
}
\description{
Is a number whole?
}
