% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_predicates.R
\name{lang-is-callable}
\alias{lang-is-callable}
\alias{is_callable}
\alias{are_callable}
\title{Is an object callable?}
\usage{
is_callable(x)

are_callable(x, list_only = FALSE)
}
\arguments{
\item{x}{an arbitrary R object.}
}
\value{
TRUE or FALSE or a vector of logicals.
}
\description{
Is an object callable?
}
