% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_predicates.R
\name{lang-is-expr}
\alias{lang-is-expr}
\alias{is_expr}
\alias{are_expr}
\title{Is an object an expression?}
\usage{
is_expr(x)

are_expr(x, list_only = FALSE)
}
\arguments{
\item{x}{an arbitrary R object.}
}
\value{
TRUE or FALSE or a vector of logicals.
}
\description{
Is an object an expression?
}
