% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_predicates.R
\name{lang-is-symbolic}
\alias{lang-is-symbolic}
\alias{is_symbolic}
\alias{are_symbolic}
\title{Is an object a call or a symbol?}
\usage{
is_symbolic(x)

are_symbolic(x, list_only = FALSE)
}
\arguments{
\item{x}{an arbitrary R object.}
}
\value{
TRUE or FALSE or a vector of logicals.
}
\description{
Is an object a call or a symbol?
}
