% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_predicates.R
\name{lang-is-syntactic-literal}
\alias{lang-is-syntactic-literal}
\alias{is_syntactic_literal}
\alias{are_syntactic_literal}
\title{Is an object a syntactic literal?}
\usage{
is_syntactic_literal(x)

are_syntactic_literal(x, list_only = FALSE)
}
\arguments{
\item{x}{an arbitrary R object.}
}
\value{
TRUE or FALSE or a vector of logicals.
}
\description{
Is an object a syntactic literal?
}
