% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_call.R
\name{match-call}
\alias{match-call}
\alias{match_call}
\alias{match_call_self}
\title{Call Matching (similar to \link[rlang:call_match]{rlang::call_match})}
\usage{
match_call(
  definition = sys.function(sys.parent()),
  call = sys.call(sys.parent()),
  defaults = FALSE,
  ignore_missing = FALSE,
  expand_dots = TRUE,
  envir = parent.frame(2L)
)

match_call_self(
  call,
  defaults = FALSE,
  ignore_missing = FALSE,
  expand_dots = TRUE,
  envir = parent.frame(2L)
)
}
\arguments{
\item{definition}{a function, by default the function from which
    \code{match.call} is called.  See details.}

\item{call}{an unevaluated call to the function specified by
    \code{definition}, as generated by \code{\link[base]{call}}.}

\item{defaults}{\code{TRUE} or \code{FALSE} (default) .If \code{TRUE}, missing arguments are matched to their defaults.}

\item{ignore_missing}{\code{TRUE} or \code{FALSE} (default). If \code{TRUE}, missing
arguments without a default value are removed from the matched call.}

\item{expand_dots}{logical. Should arguments matching ... in the call be
included or left as a ... argument?}

\item{envir}{an environment, from which the ... in call are retrieved, if
any.}
}
\value{
a call.
}
\description{
Call Matching (similar to \link[rlang:call_match]{rlang::call_match})
}
