% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dots.R
\name{new-dots}
\alias{new-dots}
\alias{get_dots}
\alias{dots_eval}
\alias{dots_splice}
\title{Collect, evaluate, and/or splice dots(\code{...})}
\usage{
get_dots()

dots_eval(..., .splice = FALSE, .env = parent.frame())

dots_splice(..., .env = parent.frame())
}
\arguments{
\item{.splice}{.}

\item{.env}{.}

\item{.eval}{.}
}
\value{
a list.
}
\description{
Collect, evaluate, and/or splice dots(\code{...})
}
