% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\name{ns_get_fns}
\alias{ns_get_fns}
\title{Get Namespace Functions}
\usage{
ns_get_fns(x, exports_only = NULL)
}
\arguments{
\item{x}{A function, a string, or an environment.}

\item{exports_only}{If \code{TRUE} (default), it returns only the exported functions. If \code{FALSE}, the private ones are included too.}
}
\value{
A named list of functions.
}
\description{
Get Namespace Functions
}
\examples{
ns_get_fns(mean)
ns_get_fns(environment(mean))

export_fns <- ns_get_fns("base")
all_fns <- ns_get_fns("base", exports_only = FALSE)
length(export_fns)
length(all)
}
