% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-purrr.R
\name{purrr-standalone}
\alias{purrr-standalone}
\alias{map}
\alias{walk}
\alias{map_lgl}
\alias{map_int}
\alias{map_dbl}
\alias{map_chr}
\alias{map2}
\alias{map2_lgl}
\alias{map2_int}
\alias{map2_dbl}
\alias{map2_chr}
\alias{imap}
\alias{pmap}
\alias{keep}
\alias{discard}
\alias{map_if}
\alias{compact}
\alias{transpose}
\alias{list_t}
\alias{every}
\alias{some}
\alias{negate}
\alias{reduce}
\alias{reduce_right}
\alias{accumulate}
\alias{accumulate_right}
\alias{detect}
\alias{detect_index}
\title{\code{purrr} Standalone Replacement Functions}
\usage{
map(.x, .f, ...)

walk(.x, .f, ...)

map_lgl(.x, .f, ...)

map_int(.x, .f, ...)

map_dbl(.x, .f, ...)

map_chr(.x, .f, ...)

map2(.x, .y, .f, ...)

map2_lgl(.x, .y, .f, ...)

map2_int(.x, .y, .f, ...)

map2_dbl(.x, .y, .f, ...)

map2_chr(.x, .y, .f, ...)

imap(.x, .f, ..., .index_only = FALSE)

pmap(.l, .f, ...)

keep(.x, .f, ...)

discard(.x, .p, ...)

map_if(.x, .p, .f, ...)

compact(.x)

transpose(.l)

list_t(.l)

every(.x, .p, ...)

some(.x, .p, ...)

negate(.p)

reduce(.x, .f, ..., .init)

reduce_right(.x, .f, ..., .init)

accumulate(.x, .f, ..., .init)

accumulate_right(.x, .f, ..., .init)

detect(.x, .f, ..., .right = FALSE, .p = is_true)

detect_index(.x, .f, ..., .right = FALSE, .p = is_true)
}
\description{
\code{purrr} Standalone Replacement Functions
}
