% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{set-names}
\alias{set-names}
\alias{set_names}
\alias{new_names}
\alias{as_named}
\title{Naming Helpers}
\usage{
set_names(x, nm = x)

new_names(x, pos, nm)

as_named(x)
}
\arguments{
\item{x}{a vector.}

\item{nm}{a character vector or a string.}
}
\value{
\code{set_names} and \code{as_named} return the renamed input vector. \code{is_named0} and \code{is_named} return a single \code{TRUE} or \code{FALSE}. \verb{are_*} return a logical vector of the same length as the input. \verb{which_*} return a vector of integers as postions (could be length 0).
}
\description{
Naming Helpers
}
