% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_line}
\alias{str_line}
\title{String Lines}
\usage{
str_line(..., .f = NULL, .args = NULL, .trailing = FALSE)
}
\arguments{
\item{...}{extra arguments for \code{.f}.}

\item{.f}{a function or a formula.}

\item{.trailing}{TRUE or FALSE (default).}

\item{x}{a character vector.}
}
\value{
a string.
}
\description{
String Lines
}
\examples{
out1 <- str_line(letters[1:4])
out2 <- str_line(letters[1:4], str_parens)
out3 <- str_line(letters[1:4], str_parens, .trailing = TRUE)
cat(out1, "else")
cat(out2, "else")
cat(out3, "else")
}
