% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_oxford}
\alias{str_oxford}
\title{Collapse a character vector using Oxford Comma}
\usage{
str_oxford(x, or = FALSE)
}
\arguments{
\item{x}{a character vector.}

\item{or}{TRUE or FALSE. If TRUE, the last element is coillapsed using 'or'.
If FALSE (default), the character is collapsed using 'and'.}
}
\value{
a string.
}
\description{
Collapse a character vector using Oxford Comma
}
\examples{
str_oxford(letters[1:2])

str_oxford(letters[1:4])

str_oxford(letters[1:4], or = TRUE)

}
