% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string-collapse}
\alias{string-collapse}
\alias{str_collapse}
\alias{str_coll}
\alias{str_p0}
\alias{str_p}
\title{Collapse Strings}
\usage{
str_collapse(x, .c = "", ..., collapse = .c, recycle0 = FALSE)

str_coll(x, .c = "", ..., collapse = .c, recycle0 = FALSE)

str_p0(x, .c = "", ..., collapse = .c, recycle0 = FALSE)

str_p(x, .c = " ", ..., collapse = .c, recycle0 = FALSE)
}
\arguments{
\item{x}{a character vector.}

\item{.c}{collapser. A string, a function, or a formula. It defines how
strings in the vector are collapsed. If NULL (default), the embracing gets
applied to each element of the input.}

\item{...}{optional arguments of \code{.c}.}

\item{collapse}{defaults to \code{.c}.}

\item{recycle0}{logical indicating if zero-length character arguments should
result in the zero-length \code{character(0)}. When \code{collapse} is a string,
\code{recycle0} does not recycle to zero-length, but to \code{""}.}

\item{.trailing}{TRUE or FALSE. If TRUE (default), there is no return
(\code{"\\n"}) at the end of the output. If FALSE, a return (\code{"\\n"}) is pasted at
the end of the output.}
}
\value{
a string.
}
\description{
Collapse Strings
}
\examples{
str_collapse(letters)
str_collapse(letters, .c = str_oxford)
}
