% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string-embrace}
\alias{string-embrace}
\alias{str_embrace}
\alias{str_parens}
\title{Embrace Strings}
\usage{
str_embrace(x, left, right = left, .c = NULL, ...)

str_parens(x, type = c("round", "squared", "curly", "angle"), .c = NULL, ...)
}
\arguments{
\item{x}{a character vector or an object coercible to one.}

\item{left, right}{a string.}

\item{.c}{collapser. A string, a function, or a formula. It defines how strings in the vector are collapsed. If NULL (default), the embracing gets applied to each element of the input.}

\item{...}{optional arguments of \code{.c}.}

\item{type}{one of \code{c("round", "squared", "curly", "angle")}. The type of parentheses.}
}
\value{
a string or a character vector.
}
\description{
Embrace Strings
}
\examples{
x <- gsub("\\\\s","_",rownames(mtcars))[1:12]
x

str_embrace(x, left = "-")

str_embrace(x, left = "-", right = "--")

str_embrace(x, left = "-", .c = str_oxford)

str_parens(x)

str_parens(x, "round")

str_parens(x, "round", .c = str_oxford)

str_parens(x, "squared")

str_parens(x, "curly")

str_parens(x, "angle")
}
