% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string-itemize}
\alias{string-itemize}
\alias{str_enum}
\alias{str_itemize}
\title{Enumerate and format the elements of a character vector.}
\usage{
str_enum(x, sep = ": ", pad_num = TRUE, format_str = TRUE)

str_itemize(
  x,
  type = c("list", "enum", "letters", "up_letters"),
  sep = "-",
  format_str = TRUE,
  ...
)
}
\arguments{
\item{x}{a character vector.}

\item{sep}{a string. For \link{str_enum} it indicates the separator betweem item numbers and
strings. For, \link{str_itemize}, it represents the item indicator.}

\item{pad_num}{TRUE or FALSE. If TRUE (default), numbers are padded with
str_pad_num.}

\item{format_str}{TRUE or FALSE. If TRUE (default), strings are formatted
with str_format.}
}
\value{
a character vector.
}
\description{
Enumerate and format the elements of a character vector.
}
\examples{
x <- gsub("\\\\s", "_", rownames(mtcars))[1:12]
str_enum(x)
# different separator
str_enum(x, sep = "- ")
# numbers are not padded if `pad_num` is FALSE
str_enum(x, pad_num = FALSE)
# strings not of same width if `format_str` is FALSE
str_enum(x, format_str = FALSE)

item1 <- str_itemize(x)
# identical to...
item2 <- str_itemize(x, type = "list", item = "-", format_str = TRUE)
item1
item2

# str_itemize is perfect for `cat`:
cat(item1, sep = "\n")
}
