% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string-line}
\alias{string-line}
\alias{str_return}
\alias{str_ret}
\alias{cat_return}
\alias{cat_ret}
\title{String Lines}
\usage{
str_return(..., .f = NULL, .args = NULL, .trailing = FALSE)

str_ret(..., .f = NULL, .args = NULL, .trailing = FALSE)

cat_return(..., .f = NULL, .args = NULL)

cat_ret(..., .f = NULL, .args = NULL)
}
\arguments{
\item{...}{strings or R objects coercible to strings.}

\item{.f}{a function or a formula. It uses a simplified version of \link[rlang:as_function]{rlang::as_function}.}

\item{.args}{list with extra arguments for \code{.f}.}

\item{.trailing}{TRUE or FALSE. If TRUE (default), a return (\code{"\\n"}) is pasted at the end of the output string.}
}
\value{
a string made of input strings separated by returns (\code{"\\n"}).
}
\description{
String Lines
}
\examples{
out1 <- str_line(letters[1:4])
out1
out2 <- str_line(letters[1:4], .f = str_parens, .args = list(type = "angle"))
out2
out3 <- str_line(letters[1:4], .f = str_parens, .trailing = TRUE)
out3
cat_line0(out1, "else")
cat_line0(out2, "else")
cat_line0(out3, "else")
}
