% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string-to-symbolic}
\alias{string-to-symbolic}
\alias{str_to_symbolic}
\alias{chr_to_symbolic}
\title{Convert Strings to Symbolic Objects (generalizes \link[base:parse]{base::str2lang})}
\usage{
str_to_symbolic(x, strict = FALSE)

chr_to_symbolic(x, strict = FALSE)
}
\arguments{
\item{x}{a string or a character vector.}

\item{strict}{TRUE or FALSE (default). If TRUE, only strings are accepted as
input. Other objects will throw an error even if coercible to
strings/character.}
}
\value{
a symbolic object or a lsit of symbolic objects.
}
\description{
Convert Strings to Symbolic Objects (generalizes \link[base:parse]{base::str2lang})
}
\examples{
str_to_symbolic("mean")
str_to_symbolic("mean(1:10)")
str_to_symbolic("+")
chr_to_symbolic(c("a", "c", "d", "e"))
}
