% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_to_strings.R
\name{symbolics-to-strings}
\alias{symbolics-to-strings}
\alias{lang_to_str}
\alias{lang_to_chr}
\title{Expressions to Strings or Character Vectors}
\usage{
lang_to_str(x, .names_type = c("full", "sym", "head"))

lang_to_chr(x, .names_type = c("full", "sym", "head"))
}
\arguments{
\item{x}{a symbol, or a call.}

\item{.names_type}{one of \code{c("full", "sym", "head")}.
\describe{
\item{"full"}{The output is equivalent to \code{deparse(x)}.}
\item{"sym"}{The function returns the string equivalent of a symbol \code{x} is a symbol or a call with no arguments.}
\item{"head"}{The function returns only the string equivalent of a symbol or the call name.}
}}
}
\value{
a string for \code{symbolic_to_str} or a character vector for
\code{symbolic_to_chr}.
}
\description{
Expressions to Strings or Character Vectors
}
\examples{
symbolics <- list(quote(mean), quote(mean()), quote(mean(1:10)))

symbolic_to_str(symbolics[[1]])
symbolic_to_str(symbolics[[2]])
symbolic_to_str(symbolics[[3]])

# Different `.names_type`
symbolic_to_str(symbolics[[2]], .names_type = "full")

symbolic_to_str(symbolics[[2]], .names_type = "sym")
symbolic_to_str(symbolics[[3]], .names_type = "sym")

symbolic_to_str(symbolics[[3]], .names_type = "head")

# `symbolic_to_chr` converts a list of symbolics in a vector of characters.
symbolic_to_chr(symbolics)
}
