% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_predicates.R
\name{vectorised-predicates}
\alias{vectorised-predicates}
\alias{all2}
\alias{any2}
\alias{not_all}
\alias{not_any}
\title{Extensions of [base::all()] and [base::any()] plus their negations.}
\usage{
all2(.x, .p = identity, ...)

any2(.x, .p = identity, ...)

not_all(.x, .p = identity, ...)

not_any(.x, .p = identity, ...)
}
\arguments{
\item{.x}{an R object (ideally a list or vector).}

\item{.p}{a function call, a formula, or a string. It uses a simplified version of [rlang::as_function()].}

\item{...}{additional arguments.}
}
\value{
TRUE or FALSE.
}
\description{
Extensions of [base::all()] and [base::any()] plus their negations.
}
