% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sym.R
\name{as_sym}
\alias{as_sym}
\alias{as_syms}
\title{Convert to Symbol(s)}
\usage{
as_sym(x, .strict = FALSE, .simplify = FALSE, .no_ns = FALSE, .no_args = TRUE)

as_syms(
  ...,
  .x = NULL,
  .named = NULL,
  .strict = FALSE,
  .simplify = TRUE,
  .no_ns = FALSE,
  .no_args = FALSE
)
}
\arguments{
\item{x, .x}{.}

\item{.strict}{.}

\item{.simplify}{.}

\item{.no_ns}{.}

\item{.no_args}{.}

\item{...}{.}
}
\value{
TODO.
}
\description{
Convert to Symbol(s)
}
