% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encall.R
\name{encall}
\alias{encall}
\alias{encalls}
\title{Convert Object(s) to Call(s)}
\usage{
encall(x, .strict = FALSE, .simplify = FALSE, .no_ns = FALSE, .no_args = FALSE)

encalls(
  ...,
  .x = NULL,
  .out_type = NULL,
  .strict = FALSE,
  .simplify = FALSE,
  .no_ns = FALSE,
  .no_args = FALSE
)
}
\arguments{
\item{x, .x}{.}

\item{.strict}{.}

\item{.simplify}{.}

\item{.no_ns}{.}

\item{.no_args}{.}

\item{...}{.}

\item{.out_type}{.}
}
\value{
TODO.
}
\description{
Convert Object(s) to Call(s)
}
