% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encall.R
\name{encall}
\alias{encall}
\alias{encalls}
\title{Convert to Named Call(s)}
\usage{
encall(x, .strict = FALSE, .simplify = FALSE, .no_ns = FALSE, .no_args = FALSE)

encalls(
  ...,
  .x = NULL,
  .named = NULL,
  .strict = FALSE,
  .simplify = FALSE,
  .no_ns = FALSE,
  .no_args = FALSE
)
}
\arguments{
\item{x, .x}{.}

\item{.strict}{.}

\item{.simplify}{.}

\item{.no_ns}{.}

\item{.no_args}{.}

\item{...}{.}
}
\value{
TODO.
}
\description{
Convert to Named Call(s)
}
