% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlang.R
\name{enlang}
\alias{enlang}
\alias{enlangs}
\title{Convert to \code{"symbolic"} objects (symbols and calls)}
\usage{
enlang(
  x,
  .out_type = NULL,
  .strict = FALSE,
  .simplify = FALSE,
  .no_ns = FALSE,
  .no_args = FALSE
)

enlangs(
  ...,
  .x = NULL,
  .out_type = NULL,
  .strict = FALSE,
  .simplify = FALSE,
  .no_ns = FALSE,
  .no_args = FALSE
)
}
\arguments{
\item{x}{.}

\item{.out_type}{.}

\item{.strict}{.}

\item{.simplify}{.}

\item{.no_ns}{.}

\item{.no_args}{.}

\item{...}{.}

\item{.x}{.}
}
\value{
\code{\link[=enlang]{enlang()}} returns a symbolic object. \code{\link[=enlangs]{enlangs()}} returns a list of
symbolic objects.
}
\description{
Convert to \code{"symbolic"} objects (symbols and calls)
}
