% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn-name}
\alias{fn-name}
\alias{fn_name}
\alias{fn_name2}
\title{Get Function Name from Function}
\usage{
fn_name(x, ...)

fn_name2(x, ..., with_ns = TRUE)
}
\arguments{
\item{x}{A function or list of functions (also as string or character vector).}

\item{...}{Empty dots.}
}
\value{
A string or a character vector.
}
\description{
Get Function Name from Function
}
\examples{
# fn_name("mean")
# fn_name2("mean")

}
